% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitcox.R
\name{baseline.hazard}
\alias{baseline.hazard}
\title{Baseline hazard for a Cox proportional hazards model}
\usage{
baseline.hazard(beta, event.counts, risk.set.diffs)
}
\arguments{
\item{beta}{A vector of length p, where p is the number of model parameters}

\item{event.counts}{A vector of event counts for the unique event times}

\item{risk.set.diffs}{A list of risk set differences for the event times}
}
\value{
A vector of partial derivatives for each parameter in the model
}
\description{
Baseline hazard for a Cox proportional hazards model
}
