% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitcox.R
\name{coef.fitcox}
\alias{coef.fitcox}
\title{Extract the coefficients for a fitted Cox model}
\usage{
\method{coef}{fitcox}(model)
}
\arguments{
\item{model}{A \code{fitcox} object returned by the \code{fitcox} function.}
}
\value{
A dataframe with one row per coefficient
}
\description{
Extract the coefficients for a fitted Cox model
}
\examples{
coef(fitcox(remission ~ sample, delta="censor", data=leukemia))
}
