% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitcox.R
\name{concordance.fitcox}
\alias{concordance.fitcox}
\title{Generate concordance statistics for a fitted Cox model}
\usage{
concordance.fitcox(model, newdata, responses)
}
\arguments{
\item{model}{A \code{fitcox} object returned by the \code{fitcox} function.}

\item{newdata}{A dataframe of values to generate predictions for.}

\item{responses}{The responses to compare the predicted values with}
}
\value{
A list containing:
\describe{
   \item{n.concordant}{The number of concordant pairs}
   \item{n.discordant}{The number of disccordant pairs}
   \item{c.index}{Harrel's c-index}
}
}
\description{
Generate concordance statistics for a fitted Cox model
}
\examples{
concordance.fitcox(fitcox(remission ~ sample, delta="censor", data=leukemia), leukemia, leukemia$remission)
}
