% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitcox.R
\name{cv.fitcox}
\alias{cv.fitcox}
\title{Generate cross-validation error for a fitted Cox model}
\usage{
cv.fitcox(data, model, K)
}
\arguments{
\item{data}{A dataframe of values containing predictors and response.}

\item{model}{A \code{fitcox} object returned by the \code{fitcox} function.}

\item{K}{The number of groups to use for cross-validation}
}
\value{
The mean concordance index generated using cross-validated
}
\description{
Generate cross-validation error for a fitted Cox model
}
\examples{
cv.fitcox(leukemia, fitcox(remission ~ sample, delta="censor", data=leukemia), 5)
}
