% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data1.R
\docType{data}
\name{data1}
\alias{data1}
\title{Simple data set for testing Kaplan-Meier survival curves}
\format{
An object of class \code{data.frame} with 6 rows and 2 columns.
}
\usage{
data(data1)
}
\description{
Simple data set for testing Kaplan-Meier survival curves
}
\keyword{datasets}
