% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitkm.R
\name{fitkm.curve}
\alias{fitkm.curve}
\title{Fitted Kaplan-Meier survival curve object}
\usage{
fitkm.curve(n, level, time, n.events, n.risk, surv.prob, censored)
}
\arguments{
\item{n}{the number of observations}

\item{level}{the factor level that the survival curve has been generated for}

\item{time}{t(i), the unique failure times sorted in ascending order}

\item{n.events}{m(i) , the number of failure times at each t(i)}

\item{n.risk}{r(i), the number in the risk set at each t(i)}

\item{surv.prob}{F(i), the estimated survival function at each t(i)}

\item{censored}{the unique censoring times sorted in ascending order}
}
\description{
A survival curve object containing the results of fitting a Kaplan-Meier
survival curve. The notation corresponds to that on page 188 of the Cox 1972
paper
}
\seealso{
\code{\link{fitkm}}
}
