% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitkm.R
\name{fitkm}
\alias{fitkm}
\title{Kaplan-Meier product limit estimator}
\usage{
fitkm(formula, delta, data)
}
\arguments{
\item{formula}{A formula object with the time variable on the left-hand side.}

\item{delta}{Censoring indicator, 0 = right-censored, 1 = event occurred.}

\item{data}{A data frame containing the variables referred to in \code{formula}.}
}
\value{
An object of class \code{fitkm}, containing one or more
    \code{\link{fitkm.curve}} objects.
}
\description{
\code{fitkm} fits a Kaplan-Meier product limit estimator.
}
\examples{
fitkm(recovery ~ 1, delta="status", data=data1)
fitkm(remission ~ sample, delta="censor", data=leukemia)
}
\seealso{
\code{\link{summary.fitkm}}, \code{\link{plot.fitkm}},
    \code{\link{fitkm.curve}}
}
