% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leukemia.R
\docType{data}
\name{leukemia}
\alias{leukemia}
\title{Leukemia data set from Cox's 1972 paper}
\format{
An object of class \code{data.frame} with 42 rows and 3 columns.
}
\usage{
data(leukemia)
}
\description{
Leukemia data set from Cox's 1972 paper
}
\keyword{datasets}
