% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitkm.R
\name{lines.fitkm.curve}
\alias{lines.fitkm.curve}
\title{Plots a line for a single fitted Kaplan-Meier curve}
\usage{
\method{lines}{fitkm.curve}(curve, censor = TRUE, ...)
}
\arguments{
\item{curve}{A \code{fitkm.curve} object returned by the \code{fitkm} function.}

\item{censor}{Plot the censored observations if \code{TRUE}}

\item{...}{Further graphical parameters}
}
\description{
Plots a line for a single fitted Kaplan-Meier curve
}
