% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitkm.R
\name{median.fitkm.curve}
\alias{median.fitkm.curve}
\title{Median of a fitted Kaplan-Meier curve}
\usage{
\method{median}{fitkm.curve}(curve)
}
\arguments{
\item{curve}{A \code{fitkm.curve} object returned by \code{fitkm}.}
}
\value{
The median value for the curve.
}
\description{
Median of a fitted Kaplan-Meier curve
}
