% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitcox.R
\name{partial.log.likelihood}
\alias{partial.log.likelihood}
\title{Log partial likelihood for Cox model using Efron's method for ties}
\usage{
partial.log.likelihood(beta, event.sets, risk.set.diffs)
}
\arguments{
\item{beta}{A vector of length p, where p is the number of parameters in the model}

\item{event.sets}{A list of model matrices corresponding to the event times}

\item{risk.set.diffs}{A list of risk set differences corresponding to the event times}
}
\value{
The log partial likelihood, a scalar
}
\description{
Log partial likelihood for Cox model using Efron's method for ties
}
