% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitcox.R
\name{partial.score}
\alias{partial.score}
\title{Score function for Cox model using Efron's method for ties}
\usage{
partial.score(beta, event.sets, risk.set.diffs)
}
\arguments{
\item{beta}{A vector of length p, where p is the number of model parameters}

\item{event.sets}{A list of model matrices corresponding to the event times}

\item{risk.set.diffs}{A list of risk set differences for the event times}
}
\value{
A vector of partial derivatives for each parameter in the model
}
\description{
Score function for Cox model using Efron's method for ties
}
