% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitkm.R
\name{plot.fitkm}
\alias{plot.fitkm}
\title{Produces a plot for a set of one or more Kaplan-Meier curves}
\usage{
\method{plot}{fitkm}(model, xlab = "Time", ylab = "Survival Probability", legend.cex = 1, ...)
}
\arguments{
\item{model}{A \code{fitkm} object returned by the \code{\link{fitkm}} function.}

\item{xlab}{String for the x-axis label}

\item{ylab}{String for the y-axis label}

\item{legend.cex}{Font size for the legend}

\item{...}{Further graphical parameters to be passed to \code{plot}}
}
\value{
An object of class \code{summary.fitkm}.
}
\description{
Produces a plot for a set of one or more Kaplan-Meier curves
}
\examples{
plot(fitkm(recovery ~ 1, delta="status", data=data1))
plot(fitkm(remission ~ sample, delta="censor", data=leukemia))
}
