% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitcox.R
\name{print.fitcox}
\alias{print.fitcox}
\title{Print details of a fitted Cox model}
\usage{
\method{print}{fitcox}(model)
}
\arguments{
\item{model}{A \code{fitcox} object returned by the \code{fitcox} function.}
}
\value{
The object of class \code{fitcox}
}
\description{
Print details of a fitted Cox model
}
\examples{
print(fitcox(remission ~ sample, delta="censor", data=leukemia))
}
