% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitkm.R
\name{print.fitkm.curve}
\alias{print.fitkm.curve}
\title{Print details of a fitted Kaplan-Meier curve}
\usage{
\method{print}{fitkm.curve}(curve)
}
\arguments{
\item{curve}{A \code{fitkm.curve} object returned by the \code{fitkm} function.}
}
\value{
An object of class \code{summary.fitkm}.
}
\description{
Print details of a fitted Kaplan-Meier curve
}
