% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitkm.R
\name{print.fitkm}
\alias{print.fitkm}
\title{Print details of a set of one or more Kaplan-Meier curves}
\usage{
\method{print}{fitkm}(model)
}
\arguments{
\item{model}{A \code{fitkm} object returned by the \code{\link{fitkm}} function.}
}
\value{
The object of class \code{fitkm}.
}
\description{
Print details of a set of one or more Kaplan-Meier curves
}
\examples{
print(fitkm(recovery ~ 1, delta="status", data=data1))
}
