% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitcox.R
\name{print.summary.fitcox}
\alias{print.summary.fitcox}
\title{Print a summary for a fitted Cox model}
\usage{
\method{print}{summary.fitcox}(summary)
}
\arguments{
\item{summary}{A \code{summary.fitcox} object returned by the \code{summary.fitcox} function.}
}
\value{
The object of class \code{summary.fitcox}.
}
\description{
Print a summary for a fitted Cox model
}
\examples{
print(summary(fitcox(remission ~ sample, delta="censor", data=leukemia)))
}
