% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitcox.R
\name{summary.fitcox}
\alias{summary.fitcox}
\title{Generate a summary for a fitted Cox model}
\usage{
\method{summary}{fitcox}(model)
}
\arguments{
\item{model}{A \code{fitcox} object returned by the \code{fitcox} function.}
}
\value{
An object of class \code{summary.fitcox}.
}
\description{
Generate a summary for a fitted Cox model
}
\examples{
summary(fitcox(remission ~ sample, delta="censor", data=leukemia))
}
