% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitkm.R
\name{summary.fitkm.curve}
\alias{summary.fitkm.curve}
\title{Produces a summary for a fitted Kaplan-Meier curve}
\usage{
\method{summary}{fitkm.curve}(curve)
}
\arguments{
\item{curve}{A \code{fitkm.curve} object returned by \code{\link{fitkm}}.}
}
\value{
A dataframe containing summary information for each event time.
}
\description{
Produces a summary for a fitted Kaplan-Meier curve
}
\examples{
summary(fitkm(recovery ~ 1, delta="status", data=data1))
}
