% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitkm.R
\name{summary.fitkm}
\alias{summary.fitkm}
\title{Summarize a set of one or more Kaplan-Meier curves}
\usage{
\method{summary}{fitkm}(model)
}
\arguments{
\item{model}{A \code{fitkm} object returned by \code{\link{fitkm}}.}
}
\value{
A list of summaries for each \code{fitkm.curve} object
}
\description{
Summarize a set of one or more Kaplan-Meier curves
}
\examples{
summary(fitkm(recovery ~ 1, delta="status", data=data1))
}
