% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitcox.R
\name{fitcox}
\alias{fitcox}
\title{Cox proportional hazards model}
\usage{
fitcox(formula, delta, data)
}
\arguments{
\item{formula}{A formula object.}

\item{delta}{Censoring indicator.}

\item{data}{A data frame.}
}
\value{
An object of class \code{fitcox}.
}
\description{
\code{fitcox} fits a Cox proportional hazards model.
}
\examples{
fitcox(remission ~ sample, delta="censor", data=leukemia)
}
\seealso{
\code{\link{summary.fitcox}}
}
