% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitcox.R
\name{predict.fitcox}
\alias{predict.fitcox}
\title{Generate predictions for a fitted Cox model}
\usage{
\method{predict}{fitcox}(model, newdata)
}
\arguments{
\item{model}{A \code{fitcox} object returned by the \code{fitcox} function.}

\item{newdata}{A dataframe of values to generate predictions for.}
}
\value{
A vector of predictions for the new data
}
\description{
Generate predictions for a fitted Cox model
}
\examples{
predict(fitcox(remission ~ sample, delta="censor", data=leukemia), newdata=leukemia)
}
